Add block and approve to api

+++ src/Api/Controllers/PersonController.cs
@@ -1,9 +1,9 @@
 ﻿using Api.ViewModels.Person;
+using CrossCutting.Api.RequestResults;
 using Domain.Entities;
 using Domain.Repositories;
 using Microsoft.AspNet.Mvc;
 using System;
-using System.Collections.Generic;
 using System.Net;
 using System.Threading.Tasks;
 
@@ -19,21 +19,22 @@ namespace Api.Controllers
             this.personRepository = personRepository;
         }
 
-        // GET: api/values
         [HttpGet]
-        public async Task<IEnumerable<Person>> Get()
+        public async Task<IActionResult> Get()
         {
-            return await personRepository.Get();
+            return new HttpOkObjectResult(await personRepository.Get());
         }
 
-        // GET: api/{id}
         [HttpGet("{id:Guid}")]
-        public async Task<Person> Get(Guid id)
+        public async Task<IActionResult> Get(Guid id)
         {
-            return await personRepository.GetByIdAsync(id);
+            var person = await personRepository.GetByIdAsync(id);
+            if (person == null)
+                return new HttpNotFoundResult();
+
+            return new HttpOkObjectResult(person);
         }
 
-        // POST api/values
         [HttpPost]
         public async Task<IActionResult> Post([FromBody]AddUpdatePersonViewModel viewModel)
         {
@@ -42,22 +43,79 @@ namespace Api.Controllers
             return new HttpStatusCodeResult((int)HttpStatusCode.OK);
         }
 
-        // Put api/values
         [HttpPut]
         public async Task<IActionResult> Put([FromBody]AddUpdatePersonViewModel viewModel)
         {
             var person = await personRepository.GetByIdAsync(viewModel.Id);
+
+            if (person == null)
+                return new HttpNotFoundResult();
+
             person.ChangeName(viewModel.name);
             await personRepository.Update(person);
 
             return new HttpStatusCodeResult((int)HttpStatusCode.OK);
         }
 
-        // Delete api/{id}
         [HttpDelete("{id:Guid}")]
         public async Task<IActionResult> Delete(Guid id)
         {
-            await personRepository.Remove(id);
+            try
+            {
+                var person = await personRepository.GetByIdAsync(id);
+
+                if (person == null)
+                    return new HttpNotFoundResult();
+
+                person.Delete();
+                await personRepository.Remove(id);
+            }
+            catch (Exception ex)
+            {
+                return new BadRequestObjectResult(new AppModelStateDictionary(ex));
+            }
+
+            return new HttpStatusCodeResult((int)HttpStatusCode.OK);
+        }
+
+        [HttpPut("[action]/{id:Guid}")]
+        public async Task<IActionResult> Block(Guid id)
+        {
+            try
+            {
+                var person = await personRepository.GetByIdAsync(id);
+
+                if (person == null)
+                    return new HttpNotFoundResult();
+
+                person.Block();
+                await personRepository.Update(person);
+            }
+            catch (Exception ex)
+            {
+                return new BadRequestObjectResult(new AppModelStateDictionary(ex));
+            }
+
+            return new HttpStatusCodeResult((int)HttpStatusCode.OK);
+        }
+
+        [HttpPut("[action]/{id:Guid}")]
+        public async Task<IActionResult> Approve(Guid id)
+        {
+            try
+            {
+                var person = await personRepository.GetByIdAsync(id);
+
+                if (person == null)
+                    return new HttpNotFoundResult();
+
+                person.Approve();
+                await personRepository.Update(person);
+            }
+            catch (Exception ex)
+            {
+                return new BadRequestObjectResult(new AppModelStateDictionary(ex));
+            }
 
             return new HttpStatusCodeResult((int)HttpStatusCode.OK);
         }
